requirements.txt
fastapi>=0.110.0,<0.116.0
uvicorn[standard]>=0.27.0,<0.29.0
ultralytics>=8.3.0,<8.4.0
pillow>=10.2.0,<11.0.0
python-multipart>=0.0.9,<0.0.10
opencv-python-headless>=4.8.1,<5.0.0
aiofiles>=23.1.0,<24.0.0
torch>=2.1.0,<3.0.0
